Revert changes.
