+++ tests/cpp/metric/test_elementwise_metric.cc
@@ -8,7 +8,37 @@
 #include <memory>
 
 #include "../helpers.h"
-#include "test_elementwise_metric.h"
+
+namespace xgboost {
+namespace {
+inline void CheckDeterministicMetricElementWise(StringView name, int32_t device) {
+  auto lparam = CreateEmptyGenericParam(device);
+  std::unique_ptr<Metric> metric{Metric::Create(name.c_str(), &lparam)};
+
+  HostDeviceVector<float> predts;
+  MetaInfo info;
+  auto &h_labels = info.labels_.HostVector();
+  auto &h_predts = predts.HostVector();
+
+  SimpleLCG lcg;
+  SimpleRealUniformDistribution<float> dist{0.0f, 1.0f};
+
+  size_t n_samples = 2048;
+  h_labels.resize(n_samples);
+  h_predts.resize(n_samples);
+
+  for (size_t i = 0; i < n_samples; ++i) {
+    h_predts[i] = dist(&lcg);
+    h_labels[i] = dist(&lcg);
+  }
+
+  auto result = metric->Eval(predts, info, false);
+  for (size_t i = 0; i < 8; ++i) {
+    ASSERT_EQ(metric->Eval(predts, info, false), result);
+  }
+}
+}  // anonymous namespace
+}  // namespace xgboost
 
 TEST(Metric, DeclareUnifiedTest(RMSE)) {
   auto lparam = xgboost::CreateEmptyGenericParam(GPUIDX);
@@ -32,7 +62,7 @@ TEST(Metric, DeclareUnifiedTest(RMSE)) {
               0.6708f, 0.001f);
   delete metric;
 
-  xgboost::CheckDeterministicMetricElementWise<false>(xgboost::StringView{"rmse"}, GPUIDX);
+  xgboost::CheckDeterministicMetricElementWise(xgboost::StringView{"rmse"}, GPUIDX);
 }
 
 TEST(Metric, DeclareUnifiedTest(RMSLE)) {
@@ -57,7 +87,7 @@ TEST(Metric, DeclareUnifiedTest(RMSLE)) {
               0.2415f, 1e-4);
   delete metric;
 
-  xgboost::CheckDeterministicMetricElementWise<false>(xgboost::StringView{"rmsle"}, GPUIDX);
+  xgboost::CheckDeterministicMetricElementWise(xgboost::StringView{"rmsle"}, GPUIDX);
 }
 
 TEST(Metric, DeclareUnifiedTest(MAE)) {
@@ -82,7 +112,7 @@ TEST(Metric, DeclareUnifiedTest(MAE)) {
               0.54f, 0.001f);
   delete metric;
 
-  xgboost::CheckDeterministicMetricElementWise<false>(xgboost::StringView{"mae"}, GPUIDX);
+  xgboost::CheckDeterministicMetricElementWise(xgboost::StringView{"mae"}, GPUIDX);
 }
 
 TEST(Metric, DeclareUnifiedTest(MAPE)) {
@@ -107,7 +137,7 @@ TEST(Metric, DeclareUnifiedTest(MAPE)) {
               1.3250f, 0.001f);
   delete metric;
 
-  xgboost::CheckDeterministicMetricElementWise<false>(xgboost::StringView{"mape"}, GPUIDX);
+  xgboost::CheckDeterministicMetricElementWise(xgboost::StringView{"mape"}, GPUIDX);
 }
 
 TEST(Metric, DeclareUnifiedTest(MPHE)) {
@@ -132,7 +162,7 @@ TEST(Metric, DeclareUnifiedTest(MPHE)) {
               0.1922f, 1e-4);
   delete metric;
 
-  xgboost::CheckDeterministicMetricElementWise<false>(xgboost::StringView{"mphe"}, GPUIDX);
+  xgboost::CheckDeterministicMetricElementWise(xgboost::StringView{"mphe"}, GPUIDX);
 }
 
 TEST(Metric, DeclareUnifiedTest(LogLoss)) {
@@ -161,7 +191,7 @@ TEST(Metric, DeclareUnifiedTest(LogLoss)) {
               1.3138f, 0.001f);
   delete metric;
 
-  xgboost::CheckDeterministicMetricElementWise<false>(xgboost::StringView{"logloss"}, GPUIDX);
+  xgboost::CheckDeterministicMetricElementWise(xgboost::StringView{"logloss"}, GPUIDX);
 }
 
 TEST(Metric, DeclareUnifiedTest(Error)) {
@@ -215,7 +245,7 @@ TEST(Metric, DeclareUnifiedTest(Error)) {
               0.45f, 0.001f);
   delete metric;
 
-  xgboost::CheckDeterministicMetricElementWise<false>(xgboost::StringView{"error@0.5"}, GPUIDX);
+  xgboost::CheckDeterministicMetricElementWise(xgboost::StringView{"error@0.5"}, GPUIDX);
 }
 
 TEST(Metric, DeclareUnifiedTest(PoissionNegLogLik)) {
@@ -244,5 +274,5 @@ TEST(Metric, DeclareUnifiedTest(PoissionNegLogLik)) {
               1.5783f, 0.001f);
   delete metric;
 
-  xgboost::CheckDeterministicMetricElementWise<false>(xgboost::StringView{"mphe"}, GPUIDX);
+  xgboost::CheckDeterministicMetricElementWise(xgboost::StringView{"mphe"}, GPUIDX);
 }
